+++ src/next/middleware/http-auth.ts
@@ -1,12 +1,37 @@
 import type {NextRequest} from "next/server";
 
+export interface Credentials {
+	username: string;
+	password: string;
+}
+
 /**
  * Handles basic authentication
+ *
+ * @deprecated Use the new method 'integrateHttpBasicAuth' allowing a list of credentials to check against instead
  */
 export function handleHttpBasicAuth (
 	request: NextRequest,
 	username: string,
 	password: string,
+	responseText: string = "Auth required",
+	realmLabel: string = "Secure Area",
+) : Response | undefined
+{
+	return integrateHttpBasicAuth(
+		request,
+		[{username, password}],
+		responseText,
+		realmLabel
+	);
+}
+
+/**
+ * Handles basic authentication
+ */
+export function integrateHttpBasicAuth (
+	request: NextRequest,
+	users: Credentials[],
 	responseText: string = "Auth required",
 	realmLabel: string = "Secure Area",
 ) : Response | undefined
@@ -20,7 +45,9 @@ export function handleHttpBasicAuth (
 			.toString()
 			.split(":");
 
-		if (givenUser === username && givenPassword === password)
+		const hasValidUser = users.some(item => givenUser === item.username && givenPassword === item.password);
+
+		if (hasValidUser)
 		{
 			return;
 		}